+++ finance_toolkit/boursorama.py
@@ -6,9 +6,9 @@ from typing import Tuple
 import pandas as pd
 from pandas import DataFrame
 
-from .accounts import Account
+from .account import Account
 from .models import TxType, Configuration
-from .pipelines import Pipeline, TransactionPipeline, BalancePipeline
+from .pipeline import Pipeline, TransactionPipeline, BalancePipeline
 
 
 class BoursoramaAccount(Account):
